change error/info reporting in get_genbank_record()
The existing reporting was causing downstream issues by interfering
with stdout/potential for piping results. Stylistically, functions
that report directly to stdout/stderr are not great for producing
predictable output when used in other tools/scripts. It is better
to pass communication via a logger, that can direct output under the
control of the user.